requirements.txt
Pillow>=10.0.0
cryptography>=41.0.0
numpy>=1.24.0
scipy>=1.10.0